filter list instead of clearing and adding
The sandbox/SearchBench benchmark proves that this tactic is ideal for
nativeaot. Three options were considered:

- Rip and replace (previous choice)
- Filter with compiled expressions
- Filter with if statements

Rip and replace is fast enough, but wastes memory, and triggers more
garbage collections. Aot is 4x slower than runtime, but still fast
enough. The allocations tho.. yikes.

Compiled expressions are the fastest, and allocates half as much as rip
& replace. Unless you're nativeaot. Then the allocations are
astronomical, and 9x slower than rip & replace with the runtime. 30k gc
per 1k operations.. yikes.

Filter with if statements are the sweet spot. The runtime is slightly
faster than aot, but they're comparable. Allocations are identical. If
you're the runtime, this ifs are slightly slower than expressions, but
imperceptibly. If you're aot, it's perceptibly faster than the other two
options.

SearchBaseline now exists to track performance of the actual
MainViewModel's search, over time.